+++ js/plugins/window.js
@@ -49,6 +49,8 @@ var Window = {
         left: "auto",
         place: "auto",
         closeAction: Metro.actions.REMOVE,
+        customButtons: null,
+        clsCustomButton: "",
         onDragStart: Metro.noop,
         onDragStop: Metro.noop,
         onDragMove: Metro.noop,
@@ -208,9 +210,10 @@ var Window = {
             }
         }
 
+        buttons = $("<div>").addClass("buttons");
+        buttons.appendTo(caption);
+
         if (o.btnClose === true || o.btnMin === true || o.btnMax === true) {
-            buttons = $("<div>").addClass("buttons");
-            buttons.appendTo(caption);
 
             if (o.btnMax === true) {
                 btnMax = $("<span>").addClass("btn-max");
@@ -228,6 +231,48 @@ var Window = {
             }
         }
 
+        if (Utils.isValue(o.customButtons)) {
+            var customButtons = [];
+
+            if (Utils.isObject(o.customButtons) !== false) {
+                o.customButtons = Utils.isObject(o.customButtons);
+            }
+
+            console.log(o.customButtons);
+
+            if (typeof o.customButtons === "string" && o.customButtons.indexOf("{") > -1) {
+                customButtons = JSON.parse(o.customButtons);
+            } else if (typeof o.customButtons === "object" && Utils.objectLength(o.customButtons) > 0) {
+                customButtons = o.customButtons;
+            } else {
+                console.log("Unknown format for custom buttons");
+            }
+
+            console.log(customButtons);
+
+            $.each(customButtons, function(){
+                var item = this;
+                var customButton = $("<span>");
+
+                customButton
+                    .addClass("btn-custom")
+                    .addClass(o.clsCustomButton)
+                    .addClass(item.cls)
+                    .attr("tabindex", -1)
+                    .html(item.html);
+
+                customButton.data("action", item.onclick);
+
+                buttons.prepend(customButton);
+            });
+        }
+
+        caption.on(Metro.events.stop, ".btn-custom", function(){
+            var button = $(this);
+            var action = button.data("action");
+            Utils.exec(action, [button], this);
+        });
+
         win.attr("id", o.id === undefined ? Utils.elementId("window") : o.id);
 
         if (o.resizable === true) {
@@ -240,12 +285,15 @@ var Window = {
             that.maximized(e);
         });
         btnMax.on(Metro.events.start, function(e){
+            if (Utils.isRightMouse(e)) return;
             that.maximized(e);
         });
         btnMin.on(Metro.events.start, function(e){
+            if (Utils.isRightMouse(e)) return;
             that.minimized(e);
         });
         btnClose.on(Metro.events.start, function(e){
+            if (Utils.isRightMouse(e)) return;
             that.close(e);
         });