+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + [x] Window: fix system button click behavior when draggable enabled
 + [x] Window: fix creating `icon` and `title` if these not defined
 + [x] ListView: add trigger `change` when nodes selected/deselected, issue #1313
++ [ ] Window: add custom buttons to caption
 
 ### 4.2.35
 + [x] Draggable: refactoring